Influencer points updating with an order but hardcoded with 10% of the order value

+++ apps/app/lib/app/contracts/contract.ex
@@ -3,11 +3,11 @@ defmodule App.Contracts.Contract do
   import Ecto.Changeset
 
   schema "contracts" do
-    belongs_to :brand, App.Brands.Brand
-    belongs_to :influencer, App.Influencers.Influencer
-    field :minimum_points, :integer, default: 0
-    field :payment_period, :integer, default: 0
-    field :points, :decimal, default: 0
+    belongs_to(:brand, App.Brands.Brand)
+    belongs_to(:influencer, App.Influencers.Influencer)
+    field(:minimum_points, :integer, default: 0)
+    field(:payment_period, :integer, default: 0)
+    field(:points, :integer, default: 0)
     has_many(:voucher, App.Vouchers.Voucher)
 
     timestamps()
@@ -20,6 +20,10 @@ defmodule App.Contracts.Contract do
     |> cast_assoc(:brand)
     |> cast_assoc(:influencer)
     |> validate_required([:influencer_id, :brand_id])
-    |> unique_constraint(:unique_index_brand_influencer, name: :index_brand_influencer, message: "A tua mae")
+    |> unique_constraint(
+      :unique_index_brand_influencer,
+      name: :index_brand_influencer,
+      message: "A tua mae"
+    )
   end
 end
